feat: more slimdown of helapordo.h
