+++ src/game_utils.c
@@ -2875,3 +2875,466 @@ void resetArtifactsState(Fighter *f)
         };
     };
 }
+
+/**
+ * Takes a Enemy pointer and two integers for current room index and current enemy index, and prints the spawn message formatted.
+ * @see Enemy
+ * @see stringFromEClass()
+ * @param e The Enemy pointer to print.
+ * @param roomIndex The index of current room.
+ * @param enemyIndex The index of current enemy.
+ */
+void printSpawnMessage(Enemy *e, int roomIndex, int enemyIndex)
+{
+    if (!e->beast) {
+        //TODO:
+        //Where to print a windowed spawn message?
+        log_tag("debug_log.txt", "[ROOM]", "Room %i)\t\t%s\t\tEnemy #%i",
+                roomIndex, stringFromEClass(e->class), enemyIndex + 1);
+    } else {
+        log_tag("debug_log.txt", "[ROOM]",
+                "Room %i)\t\t%s\t\tEnemy #%i IS BEAST", roomIndex,
+                stringFromEClass(e->class), enemyIndex + 1);
+        //lightBlue();
+        //printf("\nYou feel at discomfort.\n\nRoom %i)\t\t",roomIndex);
+        //lightRed();
+        //printf("Enemy #%i:\t%s",enemyIndex+1,stringFromEClass(e->class));
+        //white();
+    }
+}
+
+/**
+ * Takes a Fighter pointer value and adds a random Consumable to consumablesBag.
+ * @see Fighter
+ * @see Consumable
+ * @see consumableClass
+ * @see CONSUMABLESMAX
+ * @see stringFromConsumables()
+ * @param player The Fighter pointer at hand.
+ * @return int Returns the enum value of the drop as an integer.
+ */
+int dropConsumable(Fighter *player)
+{
+    int drop = rand() % (CONSUMABLESMAX + 1);
+
+    //Special drop chances. Maybe a function for this?
+    if (drop == Powergem) {
+        if (rand() % 3 == 0) {
+            drop = rand() % (CONSUMABLESMAX + 1);
+        }
+    }
+    // Powergem has 33% chance to be rerolled
+
+    Consumable *c = (Consumable *) player->consumablesBag[drop];
+    c->qty++;
+
+    //Update stats
+    player->stats->consumablesfound++;
+
+    return drop;
+}
+
+/**
+ * Takes a Fighter pointer value and adds a random Artifact to artifactsBag.
+ * The Artifact is randomised according to ARTIFACTSMAX until one which was not found yet drops.
+ * Special Equip functions are also set up here.
+ * @see Fighter
+ * @see Artifact
+ * @see artifactClass
+ * @see stringFromArtifacts()
+ * @param player The Fighter pointer at hand.
+ * @return int Returns the enum value of the drop as an integer.
+ */
+int dropArtifact(Fighter *player)
+{
+    int drop = 0;
+    do {
+        drop = rand() % (ARTIFACTSMAX + 1);
+    } while (player->artifactsBag[drop]->qty != 0);	//We reroll to get one we don't have
+
+    player->artifactsBag[drop]->qty++;
+
+    //Update stats
+    player->stats->artifactsfound++;
+
+    return drop;
+}
+
+#ifdef HELAPORDO_CURSES_BUILD
+
+/**
+ * Takes a WINDOW pointer and prints the passed text to it ith wprintw(), before sleeping for the specified amount of milliseconds.
+ * @see handleRoom_Enemies()
+ * @see handleRoom_Boss()
+ * @param w The WINDOW pointer to print to.
+ * @param text The contents of the notification.
+ * @param time The display time in milliseconds
+ */
+void display_notification(WINDOW *w, char *text, int time)
+{
+    wprintw(w, "\n  %s", text);
+    wrefresh(w);
+    //refresh();
+    napms(time);
+}
+
+/**
+ * Takes a Equip pointer and prepares its sprite field by copying it line by line from equips_sprites, defined in sprites.h header.
+ * @see Equip
+ * @see dropEquip
+ * @see equips_sprites
+ * @param e The Equip pointer whose sprite field will be initialised.
+ */
+void setEquipSprite(Equip *e)
+{
+    if (e->class < EQUIPSMAX + 1) {
+        for (int i = 0; i < 8; i++) {
+            strcpy(e->sprite[i], equips_sprites[e->class][i]);
+        }
+    } else {
+        fprintf(stderr,
+                "[ERROR]    Unexpected equipClass in setEquipSprite().\n");
+        exit(EXIT_FAILURE);
+    }
+}
+
+/**
+ * Takes a quality value and calls the respective color function without actually printing text.
+ * @see quality
+ * @param q The quality value we want to set text color for.
+ */
+void printQualityColor(quality q)
+{
+    switch (q) {
+    case Bad: {
+        lightRed();
+    }
+    break;
+    case Average: {
+        strongWhite();
+    }
+    break;
+    case Good: {
+        lightCyan();
+    }
+    break;
+    };
+}
+
+/**
+ * Takes a Equip pointer and prints most of its values formatted.
+ * The beginning of the format assumes the caller prepended "(x" on the output before calling, where x is the index of the equip.
+ * @see Equip
+ * @see stringFromQuality()
+ * @param e The Equip pointer with stats to print.
+ */
+void printEquipStats(Equip *e)
+{
+    printf(")\t");		//This completes the print started in the calling loop, which supplies the index ATM
+
+    //Quality color
+    printQualityColor(e->qual);
+    printf("%s  ", stringFromQuality(e->qual));
+
+    printf("%s  ", e->name);
+    white();
+
+    printf("\"%s\"  (L%i)\t", e->desc, e->level);
+
+    lightCyan();
+    printf("%s  ", stringFromEquipzones(e->type));
+
+    lightGreen();
+    //Stats, will be then printed only if != 0
+    if (e->atk != 0) {
+        printf("A: %i ", e->atk);
+    };
+    if (e->def != 0) {
+        printf("D: %i ", e->def);
+    };
+    if (e->vel != 0) {
+        printf("V: %i ", e->vel);
+    };
+    if (e->enr != 0) {
+        printf("E: %i", e->enr);
+    };
+
+    printf("\n");
+    white();
+
+    //Perks, will be then printed only if perksCount != 0
+
+    for (int i = 0; i < e->perksCount; i++) {
+        lightPurple();
+        printf("\t\t%s\n", e->perks[i]->name);
+    };
+    white();
+}
+
+/**
+ * Takes a Fighter pointer value and an integer indicating if the drop was from a beast enemy, and adds a random Equip to the fighter's equipsBag.
+ * Prints notifications to the passed WINDOW pointer.
+ * The Equip dropped is initalised here, including stat variations for quality and level boosts (stat increase from base level by adding player level over EQUIPLVLBOOSTRATIO.
+ * The values of earliestBagSlot and equipsBagOccupiedSlots are managed.
+ * If equipsBag is full (EQUIPSBAGSIZE), user has to choose one Equip not currently equipped to be deleted and replaced by the new one.
+ * Special Equip functions are also set up here.
+ * @see Fighter
+ * @see Equip
+ * @see equipClass
+ * @see quality
+ * @see EQUIPLVLBOOSTRATIO
+ * @see EQUIPSBAGSIZE
+ * @see stringFromQuality()
+ * @see stringFromEquips()
+ * @param player The Fighter pointer at hand.
+ * @param beast The integer for drops coming from a beast kill if true.
+ * @param notify_win The WINDOW pointer to call display_notification() on.
+ * @param kls The Koliseo used for allocations.
+ */
+void dropEquip(Fighter *player, int beast, WINDOW *notify_win, Koliseo *kls)
+{
+
+    //Select a basic item from the list
+    int drop = rand() % (EQUIPSMAX + 1);
+    //Randomise quality
+    quality q = rand() % (QUALITIESMAX + 1);
+
+    //Prepare the item
+    kls_log(kls, "DEBUG", "Prepping dropped Equip");
+    Equip *e =
+        (Equip *) KLS_PUSH_TYPED(kls, Equip, HR_Equip, "Equip", "Equip");
+
+    //Get the base item and copy the stats to the drop
+    Equip *base = &equips[drop];
+
+    e->class = base->class;
+    e->type = base->type;
+    e->qual = q;
+
+    setEquipSprite(e);
+    strcpy(e->name, base->name);
+    strcpy(e->desc, base->desc);
+
+    e->qty = 1;
+    e->equipped = 0;
+
+    e->perksCount = 0;
+
+    //Calc drop level
+    e->level = base->level + round(player->level / EQUIPLVLBOOSTRATIO);
+
+    //Chance for better leveled item
+    if ((rand() % 8) - (player->luck / 10) <= 0) {	//Should use a defined constant
+        e->level += 1;		//At least a simple +1
+        if ((rand() % 25) - (player->luck / 10) <= 0) {	//Should use a defined constant
+            e->level += 1;	//A bonus roll for another +1
+
+        }
+    }
+
+    float boostFactor = 0.7;
+
+    float lvlBoost = boostFactor * (e->level - 1);
+
+    e->atk = round((base->atk * 1.0) + lvlBoost);
+    e->def = round((base->def * 1.0) + lvlBoost);
+    e->vel = round((base->vel * 1.0) + lvlBoost);
+    e->enr = round((base->enr * 1.0) + lvlBoost);
+
+    //Bonus stats on better quality items? Simple for now
+    //
+    if (q == Good) {
+        e->atk += (rand() % 3);	//Should use a defined constant
+        e->def += (rand() % 3);	//Should use a defined constant
+        e->vel += (rand() % 3);	//Should use a defined constant
+        e->enr += (rand() % 2);	//Should use a defined constant
+    } else if (q == Bad) {
+        e->atk -= (rand() % 3);	//Should use a defined constant
+        e->def -= (rand() % 3);	//Should use a defined constant
+        e->vel -= (rand() % 3);	//Should use a defined constant
+        e->enr -= (rand() % 2);	//Should use a defined constant
+        if (e->atk < 0) {
+            e->atk = 0;
+        };
+        if (e->def < 0) {
+            e->def = 0;
+        };
+        if (e->vel < 0) {
+            e->vel = 0;
+        };
+        if (e->enr < 0) {
+            e->enr = 0;
+        };
+    }
+    //Possible perks for the Equip
+
+    for (int i = 0; i < (EQUIPPERKSMAX); i++) {
+        int chance = 20;
+
+        if (q == Good) {
+            chance *= 1.5;
+        }
+
+        if ((rand() % 100) < chance || (beast && e->perksCount == 0)) {
+
+            e->perksCount += 1;
+
+            log_tag("debug_log.txt", "[DEBUG]",
+                    "Prepping Perk (%i) for dropped Equip)", e->perksCount);
+            kls_log(kls, "DEBUG", "Prepping Perk (%i) for dropped Equip)",
+                    e->perksCount);
+            Perk *p =
+                (Perk *) KLS_PUSH_TYPED(kls, Perk, HR_Perk, "Perk", "Perk");
+            p->class = rand() % (PERKSMAX + 1);
+            //p->name = (char*)malloc(sizeof(nameStringFromPerk(p->class)));
+            strcpy(p->name, nameStringFromPerk(p->class));
+            //p->desc = (char*)malloc(sizeof(descStringFromPerk(p->class)));
+            strcpy(p->desc, descStringFromPerk(p->class));
+            p->innerValue = 1;
+            e->perks[(e->perksCount - 1)] = p;
+        }
+    }
+
+    //Set value of how many bonuses we have. Although Equip only has ONE function pointer field for now
+    e->bonus = base->bonus;
+    //Randomise if the item will have an effect function.
+    //Not yet implemented
+    //Initialisation of function happens here
+    //
+    //e->equip_fun =    ;
+
+    //Calc cost value
+
+    float cost = 5;
+
+    cost += 2.5 * (e->qual);
+    cost += 3.5 * (e->perksCount);
+
+    cost += 7.2 * (e->level);
+
+    e->cost = floor(cost);
+
+    char msg[500];
+
+    wattron(notify_win, COLOR_PAIR(S4C_BRIGHT_YELLOW));
+    sprintf(msg, "You found %s %s!", stringFromQuality(q),
+            stringFromEquips(drop));
+    display_notification(notify_win, msg, 800);
+    wattroff(notify_win, COLOR_PAIR(S4C_BRIGHT_YELLOW));
+    log_tag("debug_log.txt", "[DEBUG-DROPS]", "Found Equip:    %s.",
+            stringFromEquips(drop));
+
+    if (player->equipsBagOccupiedSlots >= EQUIPSBAGSIZE) {	//Handle full bag by asking player if we throw something away
+        //FIXME: The handling of full bag event is a mess as it does not support curses.
+        lightRed();
+        printf("\tYour bag is full. Want to throw something away?\n");
+        white();
+        log_tag("debug_log.txt", "[DEBUG-EQUIPS]",
+                "Bag full, need to make space.\n");
+
+        for (int i = 0; i < player->equipsBagOccupiedSlots; i++) {
+            Equip *current = (Equip *) player->equipsBag[i];
+            if (current->equipped) {
+                green();
+                printf("ON  ");
+                white();
+            };
+
+            printf("(%i", i);	//This starts lines with the item index.
+            printEquipStats(current);
+        };
+
+        int selected = -1;
+        int c = -1;
+        Equip *s = NULL;
+
+        while (selected < 0 || selected >= player->equipsBagOccupiedSlots
+               || c != 1) {
+
+            c = scanf("%i", &selected);
+            int res = scanf("%*c");
+            log_tag("debug_log.txt", "[DEBUG]",
+                    "dropEquip() scanf() res was (%i)", res);
+
+            if (c == 1) {
+                s = (Equip *) player->equipsBag[selected];
+                if (s->equipped) {
+                    printf("You can't delete an equipped item!");
+                    selected = -1;
+                }
+            }
+        }
+
+        printf
+        ("\tAre you sure you want to delete %s %s ?\n\t\t[0 to confirm, 1 to go back]\n",
+         stringFromQuality(s->qual), s->name);
+
+        int n = -1;
+        c = -1;
+        while (n < 0 || c != 1) {
+            c = scanf("%i", &n);
+            int res = scanf("%*c");
+            log_tag("debug_log.txt", "[DEBUG]",
+                    "dropEquip() scanf() res was (%i)", res);
+        }
+
+        if (n != 0) {		//Abort deletion, drop will not be awared.
+
+            /*
+               int perkscount = e->perksCount;
+               if (perkscount > 0) {
+               for (int i=0; i < perkscount; i++) {
+               free(e->perks[i]);
+               }
+               }
+               free(e);
+             */
+            log_tag("debug_log.txt", "[DEBUG-EQUIPS]",
+                    "Equip was not taken.\n");
+            return;
+        };
+
+        Equip *toDelete = (Equip *) player->equipsBag[selected];
+        int perkscount = toDelete->perksCount;
+        /*
+           if (perkscount > 0) {
+           for (int i=0; i < perkscount; i++) {
+           free(toDelete->perks[i]);
+           }
+           }
+         */
+        log_tag("debug_log.txt", "[DEBUG-EQUIPS]",
+                "Equip %s (%i Perks) was taken by deleting %s.\n",
+                stringFromEquips(e->class), perkscount,
+                stringFromEquips(toDelete->class));
+        /*
+           free(toDelete);
+         */
+
+        //Add drop to player bag replacing the one at the selected index
+        player->equipsBag[selected] = (struct Equip *)e;
+
+        //Update stats
+        player->stats->equipsfound++;
+
+        napms(500);
+        return;
+    };				//End if bag is full
+
+    //Add drop to player bag
+    player->equipsBag[player->earliestBagSlot] = (struct Equip *)e;
+
+    player->earliestBagSlot += 1;	//Advance equips bage pointer
+    player->equipsBagOccupiedSlots += 1;	//Advance equips bage size counter
+
+    //Update stats
+    player->stats->equipsfound++;
+}
+
+#else
+#ifndef HELAPORDO_RAYLIB_BUILD
+#error "HELAPORDO_CURSES_BUILD and HELAPORDO_RAYLIB_BUILD are both undefined.\n"
+#else
+
+#endif // HELAPORDO_RAYLIB_BUILD
+#endif // HELAPORDO_CURSES_BUILD
